clean up average_refactor and rename to average.
